pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hft"
version = "0.1.0"
description = "Hft environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "numpy>=2.4.4",
    "pydantic>=2.0.0",
    "matplotlib",
    "jinja2",
    "python-dotenv",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m hft.server.app
server = "hft.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hft", "hft.server"]
package-dir = { "hft" = ".", "hft.server" = "server" }
